mv ListSample and ListBatch to types
