+++ src/autocast/models/encoder_decoder.py
@@ -9,7 +9,6 @@ from the_well.data.normalization import ZScoreNormalization
 from torch import nn
 from torchmetrics import Metric
 
-from autocast.data.multidataset import ListBatch
 from autocast.decoders import Decoder
 from autocast.encoders.base import Encoder, EncoderWithCond, GenericEncoder
 from autocast.metrics.utils import MetricsMixin
@@ -17,6 +16,7 @@ from autocast.models.denorm_mixin import DenormMixin
 from autocast.models.multifidelity_transformer import AttentionMixer
 from autocast.models.optimizer_mixin import OptimizerMixin
 from autocast.types import Batch, Tensor, TensorBNC, TensorBTSC
+from autocast.types.batch import ListBatch
 from autocast.types.types import TensorDBM
